+++ src/app.rs
@@ -11787,6 +11787,7 @@ Error: `{err}`"
                 gene_transcript_assay_routine: None,
                 primerbank_search_report: None,
                 external_primer_pair_import_report: None,
+                primer_variant_screen: None,
                 primer_specificity_handoff: None,
                 primer_specificity_report: None,
                 transcript_qpcr_panel: None,